refactor(core): split plugin-registry helpers + unify duplicate topo-sorts
Phase A of the plugin-registry split (WS5). Barrel-preserving — the public
module src/lib/plugin-registry.ts keeps the class, the singleton, and every
export (pluginRegistry/registerCorePlugins/isCorePlugin/CorePluginRegistration/
getPluginSkills/removePluginSkillsByPlugin), so all 55 importers and ~40 test
mock.module overlays are unchanged. The implementation moves to siblings:

- plugin-registry-types.ts — PluginState / PluginFailureState /
  PluginLoadEntry / CorePluginRegistration.
- plugin-console-capture.ts — withCapturedPluginConsole +
  createPluginScopedLogger (the AsyncLocalStorage console-capture machinery).
- plugin-activation-audit.ts — logPluginActivation (#142 layer 1).
- plugin-topo-sort.ts — ONE Kahn's-algorithm sort replacing the two
  near-identical copies (topologicalSort / topologicalSortUserPlugins). The
  three ways they differed are now options: source ('built-in'|'user'),
  failOnMissingDep (core fails the dependent; the user pass pre-filters in
  loadUserPlugins so it skips the edge), and logCycle. The unified function
  always filters the result against isFailed — a no-op for the user pass
  (cycle entries are already excluded by Kahn's), so its behavior is preserved
  exactly.

plugin-registry.ts: 1,512 -> ~1,000 lines. No behavior change.

Deferred to a follow-up (behavior-touching, own commit + boot smoke): unifying
the two activation pipelines (loadPlugin / activateUserPluginEntry — they share
a tail but diverge on import mechanism, console capture, migrations, core/user
bookkeeping, and audit-log ordering) and the physical move out of src/lib.

Verified: typecheck + lint clean; registry test 52/0; full suite 5115/0;
binary build; isolated-home boot smoke — 9 plugins activate in correct topo
order through the refactored loadPlugin->buildContext path (the lone failure is
schedule's openclaw-cron ENOENT, environmental). New direct unit test for the
unified sort: 6/0.
